move them all into one file for now. that way i can use auto complete with it
