+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "onny-utils",
-  "version": "1.0.4",
+  "version": "1.0.5",
   "description": "Utils for Onny.io",
   "main": "./index.js",
   "scripts": {
